fix: move styled component to single file
