--- src/components/ButtonGroupPicker/styled/container.js
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-const StyledContainer = styled.fieldset`
-    margin: 0;
-    padding: 0;
-    border: 0;
-    position: relative;
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-`;
-
-export default StyledContainer;
