--- src/components/ButtonGroupPicker/styled/errorText.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import ErrorText from '../../Input/styled/errorText';
-
-const StyledErrorText = styled(ErrorText)`
-    text-align: center;
-`;
-
-export default StyledErrorText;
