--- src/components/ButtonGroupPicker/styled/helpText.js
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-import { MARGIN_X_SMALL } from '../../../styles/margins';
-import { FONT_SIZE_TEXT_MEDIUM } from '../../../styles/fontSizes';
-import attachThemeAttrs from '../../../styles/helpers/attachThemeAttrs';
-
-function getAlignSelf(props) {
-    return props.alignSelf || 'center';
-}
-
-const HelpText = attachThemeAttrs(styled.span)`
-    font-size: ${FONT_SIZE_TEXT_MEDIUM};
-    margin-top: ${MARGIN_X_SMALL};
-    align-self: ${getAlignSelf};
-    text-align: ${getAlignSelf};
-    color: ${props => props.palette.text.main};
-`;
-
-export default HelpText;
